remove autotune Auto tune code wasn't working so removed it in favour of a manual tuning procedure. Reduced the window size on the time based proportional oven chamber to 500ms. Made the internal cold junction sensor report to millidegrees to stop 1 degree fluctuations. Added M155 gcode to enable auto reporting of temperature periodically- this displays all tehe need information to perform the calibration procedures.

+++ include/KTypeProbe.h
@@ -52,9 +52,9 @@ typedef int32_t milliCelcius_t;
 */
 
 typedef struct {
-  int32_t tempOffsetA = -3;
+  float tempOffsetA = -3;
   uint32_t adcA = 2022;
-  int32_t tempOffsetB = 104;
+  float tempOffsetB = 104;
   uint32_t adcB = 2716;
 } KProbeCalibration;
