+++ src/gcodes/M306.cpp
@@ -10,17 +10,17 @@ bool M306::match() {
 void M306::execute() {
     KProbeCalibration& calibration = probe.getProbeCalibration();
     if ( GCode.HasWord('A'))
-        calibration.tempOffsetA = (int)GCode.GetWordValue('A');
+        calibration.tempOffsetA = GCode.GetWordValue('A');
     if ( GCode.HasWord('B'))
         calibration.adcA = (int)GCode.GetWordValue('B');
     if ( GCode.HasWord('C'))
-        calibration.tempOffsetB = (int)GCode.GetWordValue('C');
+        calibration.tempOffsetB = GCode.GetWordValue('C');
     if ( GCode.HasWord('D'))
         calibration.adcB = (int)GCode.GetWordValue('D');
     probe.setProbeCalibration(calibration);
 
     output.printf("; Oven chamber K Type probe\r\n");
-    output.printf("   M306 A%i B%i C%i D%i\r\n",
+    output.printf("   M306 A%0.2f B%i C%0.2f D%i\r\n",
             calibration.tempOffsetA,
             calibration.adcA,
             calibration.tempOffsetB,
